- moved Note resource into own module
